fix: address review findings — exact float read-backs, precise source skip, retry guard, kind validation, TS 5.2 consumer fix
- emergencyResponse float32 fields (oomMemoryMultiplier, cpuThrottling*)
  read back via shortest-round-trip conversion like every other float — the
  raw float64 cast produced 0.800000011920929-style refresh drift.
- NodePolicy.Read skips exactly the known-virtual source ("cluster")
  instead of any unknown source, so a future backend source tag cannot
  silently orphan a managed policy.
- retryInterceptor passes straight through when maxAttempts < 1 instead of
  returning (nil, nil).
- workloadRule.kind validated against the supported kinds
  (Deployment/StatefulSet/DaemonSet/CronJob/Job) — the API accepts any
  string and a typo would create a rule that never matches.
- Stale cluster token docs removed (Describe + Update comment still said
  tokens rotate on import).
- NodeJS consumer test pins @types/node to ^18 via npm overrides: the
  latest @pulumi/pulumi pulls an @types/node that needs TS >= 5.6, which
  has been failing the TS 5.2 matrix on main's nightly since Aug 9.

+++ provider/pkg/resources/workload_rule.go
@@ -877,12 +877,12 @@ func emergencyResponseFromProto(p *apiv1.EmergencyResponseConfig) *EmergencyResp
 	}
 	return &EmergencyResponseConfigArgs{
 		OomEnabled:              p.OomEnabled,
-		OomMemoryMultiplier:     float64(p.OomMemoryMultiplier),
+		OomMemoryMultiplier:     f32(p.OomMemoryMultiplier),
 		OomMaxReactions:         int(p.OomMaxReactions),
 		OomCooldownSeconds:      int(p.OomCooldownSeconds),
 		CpuThrottlingEnabled:    p.CpuThrottlingEnabled,
-		CpuThrottlingThreshold:  float64(p.CpuThrottlingThreshold),
-		CpuThrottlingMultiplier: float64(p.CpuThrottlingMultiplier),
+		CpuThrottlingThreshold:  f32(p.CpuThrottlingThreshold),
+		CpuThrottlingMultiplier: f32(p.CpuThrottlingMultiplier),
 	}
 }
